adapted to change in types on background shift

+++ backgrounds/AberhartBackground.java
@@ -88,12 +88,12 @@ public class AberhartBackground implements Background {
 	}
 
 	@Override
-	public void setShiftX(int shiftX) {
+	public void setShiftX(double shiftX) {
 		//ignore
 	}
 
 	@Override
-	public void setShiftY(int shiftY) {
+	public void setShiftY(double shiftY) {
 		//ignore
 	}
